Bug#13960580 SUBQUERY MATERIALIZATION IS TOO RESTRICTIVE ON DATA TYPES
Subquery materialization used to require inner and outer types to match,
now let's allow _some_ cases of difference:
- ok if inner is numeric (as there's a well-defined way to cast the
outer to a number)
- otherwise, require same collation, and:
- ok if inner is temporal (as there's a well-defined way to cast the
outer to a temporal)
- ok if two strings.

So, compared to before, we have relaxed numbers, BIT and temporal.
The motivation is that an INT could be compared to a SUM of INT
(which is a DECIMAL) without resorting to IN-to-EXISTS.
Another motivation is to make decisions simpler to anticipate,
for the user and us devs.

A new test file is added, based on Harsha's previous patches.
It compares combinations of types
when used with an equi-join and when used with IN(subquery).
Results have been validated against trunk, we only have EXPLAIN
changes, showing that subq-mat and semijoin-mat are used more often.
One exception: compared to trunk, the queries which look up a
TIME into an indexed TIMESTAMP now give a good result, because they now
use subquery materialization, while IN-to-EXISTS has a bug (bug 75644).

opt_hints_subquery: changed test to maintain intention.

Approved by: Roy Lyseng <roy.lyseng@oracle.com>

Change-Id: Ia7af06f74ea0ca1f16040195952d7fb1fe9e4153