Issue #44 Group Commit: No need to flush to latest lsn of redo before writing binlog
During first stage of group commit, it'll invoke ha_flush_logs to make
sure that redo log is persisted inside engine before writing binlog.
In innodb, innobase_flush_logs is invoked, and the leader of group commit
will take latest lsn and wait for background thread writing up to this point.
But actually it only needs to wait for the lsn which is the biggest value of the group.