Issue #46 Support pseudo_server_id
A new session level variable pseudo_server_id is added in this commit. the
feature is ported from Percona.

when set the session variable,the session write binlog will use the variable
replace orignal server id.

The variable pseudo_server_id that defines the session scope is 0 by default.
If the variable is set no non-0,the server_id variables of this conneciton is set
to the same non-0 value,so write binlog events with the value of pseudo_server_id
as the server_id filed of the event. If pseudo_server_id = 0,the sever_id of the
current session is automaticially set to the vaue of global server_id.