Issue #47 bug87389: Replication position isn't persisted correctly for XA transactions
If an XA_PREPARE_EVENT log event is executed in SQL thread rather than its slave
worker thread, the replication positions was not flushed to table after executing
the event. We have to do so after executing the XA_PREPARE log event.

Unlike XID_EVENT, we can not update the positions in the user transaction otherwise
we would hold transactional row locks of the slave_relay_log_info table for too long
and block many other active transactions for too long.

Also, for XA_PREPARE_EVENT executed by slave worker threads, the positions are already
updated in rli object and stored to the worker's row in the slave_worker_info table,
so don't do that again for this event in the places as shown in the commit.