Issue #54 Limit the scope of releasing GAP lock in trx_prepare
In order to fix a lock confliction caused by XA transaction on slave,
it released gap locks in trx_prepare(). But actually this will affect almost
all kind of read-write workload with RC isolation and lead to higher contention
on lock_sys mutex. We can limit the scope to slave thread and the thread which
already holds gap lock.