Issue #53 Too long a lock wait can also be logged to slow query
Problem:
when an sql run quickly(1 ms),but it wait lock too more(10s), In this case,
it is not recorded to a slow query and is not easy to analyze the problem

Fix:
Add a new variable tdsql_compute_query_time_for_slow_logging. This variable decides
the start time of calculating execution time, Setting it to non-zero value will control
query time for slow log by measuring from query reception(2) or query start(1) or after
locking(0) to end time.