Issue #64 Port bugfix: fix tdsql_hide_partitions bug when the table has sub partition or not continues partition
This is a straightforward(slight changes) porting from tdsql-5.7:
https://git.code.oa.com/TXSQL/Percona/commit/eb6a3762df09dab8e714a3184453bbd569d9959d

Before this commit, it uses a start part id to calculate partition number
according to index, but it leads to wrong value if there's subpartition or
paritition id is not consecutive.

In this commit, all partition ids are stored in a vector. So we can get correct
id according to bitmap index