Issue #73 Reduce CPU cost of ha_innopart::clear_blob_heaps
when there's many partitions on the table, it iterates all partitions
to check if lob heap isn't null and free them. If blob is not used, then
it should do nothing. but the iteration itself costs too much cpu. In
the benchmark, it costs 0.44% cpu.

We can add a flag to indicate if there's any blob heap used. if not,
then skip the checking.