[perf] issue#20 optimization for generating unique binlog file name
LOCK_LOG will be held in the whole process of generating a
new binlog file name, and this could cause performance issue
where there're lots of binlog files, since this lock also
required in writing/reading binlog.
We optimized this by remember the largest number of binlog file name,
and use it to generate the new binlog file name.