Issue #77 Reduce the overhead of taking snapshot
In this commit, we removed the read view list, added state of each
read view. For all transactions, the read view object is always created
while the trx_t is created and destroyed whie free it.

To avoid unnecessary interating of lf_hash, it stores the latest read
view to a global position(trx_sys_t::mvcc), with a flag which indicates
if the global view is valid or not. it'll set to invalid when register or
unregister a read-write transaction. With such strategy, we reduce the cpu
cost of iterating lf_hash(This behavior can be controlled by new option
innodb_use_cloned_view which is true by default. )

Besides, the bottleneck of trx_sys->mutex is completely removed from user code
path. Insteadly, a new mutex trx_t::view_mutex is added to protect the content
of trx_t::read_view. So obviously whie the purge thread want to take a snapshot
for purge, it will iterate the mysql_trx_list and hole trx_t::view_mutex to copy
the trx's view.

Besides+1, A new option innodb_strict_gtid_commit is added. If setting to true(by
default), It'll write gtid to clone gtid persister, otherwise only when the thread
is apply thread or it sets gtid_next explicitly, it'll do the job. According to
result it looks fine if we turn it off. It's possiblely because the gtid set is
written to table on rotation of binlog. But let's add this option to avoid corner
cases we don't aware.