Issue #83 Allow to configure reserved space of index page
Currently the innodb page always reserve 1/16 space. But if
we know that the workload is almost in-place update, then it
should be better to reserve less space. This commit will add a
new option to control the factor.

New variable: innodb_page_reserve_factor
default value: 16
range: 2 - 16384
It can be changed on fly, and by default 16 means 1/16 of the page
is reserved.