Issue #95 InnoDB unnecessarily writes unmodified pages
While committing mini transaction, it'll always add x/sx locked
pages to flush list even though they are not modified. This will
lead to unnecessary data writting.

This commit add a new option innodb_mtr_check_dirty, by default it's
off which is the original logic. If turning it on, it'll mark the mtr
slot as dirty before writing log header for the page. And while commtting
the mtr, the flag is checked.