when shard key is varchar,create table will fail.
The root cause was that when we create table in the shard way,
it did not belong to the built-in hash type of the system.

The system determined that the partition field was utf8 or other multi-character set, which was not allowed to happen.

Modification: for TDSQL shard tables, skip this restriction

ref issue: https://git.code.oa.com/TXSQL/MySQL-8.0/issues/98