Issue #96 Cache the result of records_in_range()
While testing tpcc, we found that the invoke of records_in_range()
leads to extra io cost because it needs to estimate records in
range. Due to the fact that the schema of prepared statement or
stored procedure is typically stable, we can resue the results
of records_in_range() in next query.

This patch specially cached the results of records_in_range() for
sp and prepared statement. By default it's not used. You can enable
it by turning on option sp_cache_range_info.

It collects 16 values and then use the average value in the following
1000000 calls. After that the cached value is cleared and re-cached.

The old option innodb_skip_dml_estimate_range and related code is
completely removed.