Issue #105 Reduce mutex contention of purge threads
When there are many purge threads, suspend and release the threads
are protected by same mutex, and this may lead to unnecessary
contention.

Actually the thread slot is inited on startup and freed on shutdown,
and using a local mutex to protet slot itself is enough.

So for reserving or freeing slot, using global srv_sys->mutex
For suspending/releasing slot, using slot->mutex