Issue #131 Performance Regression with XA transaction
While using xa trasaction, it'll add gtid to a in-memory
table and wait them to be flushed in trx_commit_for_mysql()

I didn't see any problem if skipping wait_flush(). It looks like
it's designed for clone plugin. But for now let's skip it. Also
the behavior is controled by option innodb_strict_gtid_commit.