Issue #104 binlog.binlog_xa_prepared_disconnect fails sometime
flags2_inited is not initialized to zero while creating query event
for XA_PREPARED_LIST ''. While parsing binlog, flags2_inited maybe
true and after that it tries to parse flag, but there's nothing, so
a sql is generated like this:
SET /*!*/;
This isn't a legal statement and lead to failure of case

Fix: init flags2_inited to zero.