Issue #161 Relax Privilege checking for store procedure and prepared statement
Typically we don't change privielege of the user. and
once it's changed, it takes effect immediately.

This issue addes a new option to relax the privilege
checking for store procedure: If the privileges doesn't
change and the sp is succesfully executed, next time
privilege checking will be skipped. This minior change
can help saving some cpu. But due to this change, the
side effect is that the privege changing takes effect
only after finishing sp.

For now, this option is only for TPCC testing.
so by default it's turn off