Issue #176 Server hangs on shutdown while waiting for ACK
Problem:
While a session is waiting for ACK, and issues a shutdown
command, the signal_hand thread will keep waiting all thd
to be disconnected. But the listener doesn't work for the
time and the thd will keep in queue forever.

Fix:
before stopping, pop all sessions from queue and let the
answer threads to abort all of them.