Issue #194 Server crashes in lock_trx_release_read_locks_in_x_mode
While iterating lock list of the transaction, we firstly get next_lock
on list and then check current lock. If current lock's discard flag is
set but next_lock's discard is not set, we'll using next_lock directly,
otherwise we need to iterate from list header again(because both lock and
next_lock are already removed from list).

The problem is when checking next_lock->discard, the next_lock maybe a
nullptr, so we must check if the pointer firstly and then check the flag.

modify version to R650