Issuse #197 Set trx_t::skip_persist_gtid for XA commit/rollback
If a XA transaction is opened and commit/rollback by different session,
it's possible that trx_t::skip_persist_gtid is not set. So for each xa
commit of prepared transaction, it'll invoke Clone_persist_gtid::wait_flush
to waiting background thread to flush all gtids and this seriously decrease
the TPS especially when a trx is just recovered from restarting instance where
the flag is always false.

Before we just set the flag in trx_start_low, and in this commit we should also
set the flag in innobase_commit_by_xid and innobase_rollback_by_xid