Issue #198 Optimize access to undo header page
Typically if the transaction is not that big, the header page
of undo has a big chance that still in buffer pool. So in this
commit we tracks pointer of header block. and when we need to use
the block, the pointer is used as a hint.

Note that even the hint matches the request, it still needs to validate
the block with s lock of page hash, but can avoid searching the hash