Issue #232 optimaze the slave's io thead call write for every one binlog
When I press tpcc,the slave' io thread has 100% cpu,And causes large IO writes.

It is not necessary,we can call write relaylog when response the ack or
fsync relaylog,At this case,we can decrese the write call

There is a time error at the millisecond level on the source and replication machines.
so we remove those warnings logs,like this:The replication timestamps have returned to normal values

When not every binlog write, if there is a phenomenon of incomplete binlog writing in the slave ,
by default, the SQL thread of mysql will be blocked .So add an automatic truncATE mechanism,
encounter such problems automatically put the redundant incomplete binlog truncate out,
IO thread will automatically re-pull

add variable:
txsql_slave_io_optimaze_write: turn on this future. but it is false
when default,but we will turn it on in template.cnf