[perf] issue#238 reduce ahi latch contention during index building PROBLEM
During system warmup, multiple threads on btree
search code path might be competing for the btr_search_latches
because they decide to build adaptive hash index all at the same time.
Since this is a exclusive latch, only one of the thread is
allowed to be operating on the AHI hash table
while the rest are just waiting and wasting CPU cycles.

SOLUTION:
We observe that in btr_cur_search_to_nth_level, only when the
btr_search_latches is not x-latched does the system consider
performing an adaptive hash index search. (storage/innobase/btr/btr0cur.cc:939)
Because the benefit of hash search might be offset by the time
it takes to wait for the writer to release the latch.
We might be better off using the btree search instead.

Inspired by this, we can actually apply the same idea
to the contention problem described above.
In btr_search_info_update_slow, if the reader decides
to build the hash index for the page, we postpone the building
if the btr_seach_latches is x-latched by a writer,
thereby avoiding the contention on this critical search path.
This make senses as we will be better off falling back
to the btree search instead of busy waiting for the x-latch
to be released. And the hash index will eventually be built
if the page is frequently accessed latter(which is worth the effort).