[bugfix]issue#133 hash scan incorrectly applies logs resulting in HA_ERR_KEY_NOT_FOUND. (merge request !109)
Problem:
========
In the function Rows_log_event::do_scan_and_update, when a CRC value
collision occurs in the process of next_record_scan, the value of
table->record[0] will be set to the Key of the hash table collision
element (that is, the "before image" of update_row_event in the hash
table ), resulting in the following while loop being true, and then
update_row_event is applied to a wrong row.

  "while(this->get_type_code() == UPDATE_ROWS_EVENT &&
        table->s->primary_key >= MAX_KEY &&
        (entry = m_hash.get(table, &m_cols)))"

Solution:
=========
This problem is caused by getting the value in the hash table in case of
failure. So use "entry" to indicate that the last application was
successful, and then go to get the next value in the hash table.

 "while(this->get_type_code() == UPDATE_ROWS_EVENT &&
        table->s->primary_key >= MAX_KEY &&
        entry &&
        (entry = m_hash.get(table, &m_cols)))"