[perf] issue#96 Optimization of AHI cleanup for drop table (merge request !76)
PROBLEM
With a large buffer pool and adaptive hash index(AHI) enabled,
removing entries from AHI during dropping a table could
take a very long time if a large number of pages of the
the table is indexed by AHI.

SOLUTION
Given that AHI does not improve performance in most of
the cases, we employ the following optimization:

If the table's AHI entries take a majority
(determined by a preset threshold value ranging from 0 to 1)
of the total AHI entries, we cleanup the AHI system
hash tables entirely which is fast. This could mistakenly
clean up the AHI entries of other tables which might
affect the performance of other queries. However,
since AHI is a cache, this cleanup does not
result in any correctness issues. Moreover, given that
AHI does not improve performance in most of the cases,
the benefit of such cleaning outweighs the potential
performance instability. As another optimization,
we only clean up the AHI partitions that are affected,
leaving other partitions unchanged to reduce this instability.