Issue #264 crash on shutdown with assertion TRX_SYS_ANY_ACTIVE_TRANSACTIONS() == 0
Revert e49ff52f731e8389af08f5b686ee686418761101 which is a workaround of the issue
Port from upstream 8.0.23, bellow is the commmit log:

commit 59490c6d61358504cbc46fd4f08d3a6d25e3cfe6
Author: Nik?a Skeled?ija <niksa.skeledzija@oracle.com>
Date: Wed Jul 1 11:50:44 2020 +0200

Bug #31006095 - Assertion failure: TRX_SYS_ANY_ACTIVE_TRANSACTIONS() == 0

- Cleanup in certain replicated XA transactions would fail to reattach
InnoDB trx_t backup causing a leaked trx_t object which was detected on
server shutdown by the assertion in the trx sys.
- There were several failure modes, the unflag_engine_ha_data_detached
function was implemented with internal state, it would only return true
the first time it was called after a detach. In certain code paths, we
called it without calling re-attach, which obviously leads to a faulty
state.
- An additional problem was calling reattach_engine_ha_data_to_thd
for handlerton instances registered in the transaction context. This
method relies on the transaction context to get ha_info, while in xa.cc
there was a comment stating that even if we don't have ha_info in the
transaction context we still need to reattach the transaction using
thd->rpl_reattach_engine_ha_data() which sets the detached flag
internally.
- Fixed all these issues by removing state from the method that
checks whether the state is detached, and calling
rpl_reattach_engine_ha_data whenever we need to perform reattachment.