Bug#26883680 COMMIT_ORDER_MANAGER CAN'T TERMINATE MTS WORKER PROPERLY WHEN DEADLOCK HAPPENS
[post-push]

Description
-----------
Two issues introduced by BUG#26883680 fix being addressed in this patch:

1) An assertion that is being hit when the multi-threaded applier stops due
to a binlog write error.

2) A warning suppression in
 that isn't generic
enough.

Analysis/Fix
------------
The analysis and proposed fix for each of the above issues:

1) The assertion being met at  doesn't seem to
hold anymore since it assumes the method isn't invoked outside the
execution of a  statement. This isn't true anymore since a
worker of a multi-threaded applier may rollback the transaction due to
 being exhausted and, in this case, is intended for it to
function as a  statement ().

2) The suppression includes the transaction identifier, in the case
'ANONYMOUS', meaning that it was only suited for non-gtid enable
scenarios. removed the 'ANONYMOUS' part.

RB: 25641