Issue #273 Support Waiting for ACKs from multiple slave
Before the transaction commits if it accepts at least one ack from
slave. This commit adds a new variable sqlasync_wait_n_slaves. and
it's value:
- default 1, same behavior as before
- > 1, got at least this many ack, the transaction can be finished

The format of ACK packet is changed and adds a new variable thread_id
which is thread id of dump thread on master, so that we can know ack
is from which slave. Because of this, both master and slave must upgrade
to same version

Change BinlogPosAns to BinlogPosAnsV1 because it parses upd packet with
class name. By changing class name, we can upgrade slave firstly and then
master.

a)add m_server_id for BinlogPosAns to support good compatibility

1.new master & old slave, we just set sqlasync_wait_n_slaves to 1,Otherwise it will timeout
2.old master & new slave, old master use old protocal,don't effect

b)optimaze some code for better performance