Issue #271 XA deadlock leads to hang of slave
If deadlock happens between two xa transaction and the session which
was rollback executes some statement after that, the binlog will include
the following logs:
XA START;
Statement;
Commit;

This is illegal on slave and lead to hang of replication.

Fix:
This fix is from bug#99045 of offical bug system:
The deadlock case run into function , it clears OPTION_BEGIN flag.
So the subsequent DML fails to know it belongs to XA and starts a new autocommit trx.