fix issue#300 kill slave in mts recovery cause replication start failed
In mts recovery if the trx had executed before or gtid is executed. It will be skipped and the group infomation is lost.
When the events update pos it will update the group pos and let the pos in group, If server exit here
the replication can't start again from this pos when gtid is on.