Issue #320 Multiple queries in one statement may lost commit with AFTER_SYNC mode
In dispatch_command, if command type is COM_QUERY and the statement includes multiple
quries, it'll invoke mysql_parse in loop. But with after_sync mode, it only commits
transaction before ending dispatch_command. So the commits inside the loop may get lost.

Fix:
This is a workaround. For such case, we don't use afterer_sync mode. Luckily that the usage
of such kind queries is rare.