[bugfix] Unstable performance with opening innodb_hot_update_detect This is a follow-up to issue#91.
problem:
The TPS will drop down to 0 while running sysbench update_with_index
when innodb_hot_update_detect is on.
The root cause is there's a race condition in check hot update logic.
It causes that Trx has already be added to the wait queue but it
doesn't start to waiting for the grant event. And this cause the
running trx grant the rec lock to the trx which not starts to waiting,
so that the trx can't move to running status, and all other trxs will
be stuck.

solution:
Move the logic of adding trx to the waiting queue to lock waiting
function, eliminate the race condition.