[bugfix] parse_ending_of_sp_error issue#105 (merge request !86)
The parser in kebled block eps terminating semi-colon for a
statement. When it is a DDL creating a stored procedure, the
semi-colon becomes part of the definition. Tools like mysqldump
wrap the definition within a version comment which does not support
any semi-colon-terminated complete statement, as a result, the
output could not be imported.

When defining a SP (e.g., TRIGGER, PROCEDURE, and FUNCTION),
we often set new separators to distinguish the default line
separator (semicolon ';') and whole SP terminators. Normally,
the newly defined delimiter is directly used as the ending
character of the SP.
Eg.,
  DELIMITER //
  CREATE
  TRIGGER `TG` BEFORE UPDATE ON `t1` FOR EACH ROW
  LABELTRI: BEGIN
  SET new.NAME = '1';
  END//

Even if the last line incorrectly writes one more semicolon,
the extra semicolon will be deleted when storing the SP.
Eg.,
  DELIMITER //
  ...
  ...
  END;
  //

The problem is that when the user or other DBA auxiliary tools not
only write one more semicolon, but also add additional characters
before the newly defined end separator, the SP will incorrectly
stored one more semicolon.
Eg.,
  DELIMITER //
  ...
  ...
  END;
  *//
The SP is stored as:
  DELIMITER //
  ...
  ...
  END;
At this time, when mysqldump is used, the last line will generated
"end;". Which causes errors when importing the SP.

Fix:
a) When setting the SP terminator (sp_head.cc:set_body_end), we verify
that the above error and remove the extra semicolon to correct the end
pointer.
b) When mysqldump is executed, we remove the extra semicolon when
exporting the stored SP.

port from 8022, commit ids from 8022:
a194635e4f6de2e600e1b69762fc822cc6c14f6e