ISSUE #481 CRASHED ON ROLLBACK AND #ISSUE#480 RECOVERY UNDO CRASH
Bellow is the commit log from upstream:
commit f298f8922c8acced88ebe0a48c662dfeb88bb746
Author: Mayank Prasad <mayank.prasad@oracle.com>
Date: Thu Jun 18 10:46:54 2020 +0530

Bug#31391126 : ROLLBACK AN UPDATE AFTER INSTANT DDL CAUSE CRASH

Background:
After first INSTANT columns is added to table, for any new row, there
is 1 (or 2) extra byte added in record header to keep number of fields
present in the row. NOTE : This won't affect the existing rows.

Issue:
A scenario, when records was JUST fitting on cluster index page. Even
a single byte increase in record size would cause the record BLOB to
go to external pages.

Now, during row update (which causes record to change), this extra byte
is included in size calculation and it's concluded that this record
can't fit in cluster record page. So record BLOB data is moved to
external pages while keeping reference (20 byte) on index page.

When rollback happens, we try to restore the original record which was
fitting in cluster index page. But during calculation of record size,
we forget to exclude this extra byte. Because Of which it is assumed
that original record has data on external pages (which is not true).
Thus when this data is looked for, supposedly stored on ext page,
metadata doesn't match causing an assert.

Fix :
During rollback, make sure if this record is going back to original
one which was before the INSTANT columns was added, don't include this
extra byte in size calculation.