fix issue#494 buffer log cause restart mysqld is too slow
For record all log to plugin log services, log events are record in global list and
processed by all the plugins every time. If the list is too long to process the startup
will be very slow. So we set a max length of the list and if it passed only log to file.

add the mtr of debug mode

Revert Issue