[feature]#660 Support sort split
Problem:
========
Support sort splitting for parallel execution. When execute
query with sorting, we can ensure the local order of records
through the sorting operation in each worker. In order to
ensure the global order of the results on the Gather, we need
to merge sort the results sent by the worker.

Solution:
========
[1] Support to inject exchange after sort iterator.
[2] Use PX_receiver_merge to replace PX_receiver.

MR!625 reviewed by kenkchen, maxmqiu