[bugfix] bugfix_parallel_ddl_wrong_data_with_one_thread
This is a follow-up of issue#64.

Problem:
When the parallel_read_thread is greater than one but
innodb_parallel_ddl_thread = 1, using parallel ddl to create index
leads to wrong data.

This is because parallel ddl decides not to sort merge buffer in the first
scan stage to optimize performance, while it is not safe for the single-
thread merge sort stage.

Fix:
In the first scan stage, if parallel_sort_threads > 1, we do not sort
merge buffer and write it to files directly. If parallel_sort_threads = 1,
we sort merge buffer. The subsequent parallel_sort_threads may be adjuested
due to small scanned merge file. In this case, we adjust parallel_sort_threads
to merge_file.offset. In extreme cases, parallel_sort_threads may be adjusted to
1. We still perform parallel ddl with one thread instead of the original
single-threaded merge sort (since merge buffer is not sorted at all).