[feature]#613 Support extracting and reusing optimization contexts (1/3)
Problem:
========
Parallel execution works by running certain fragments of the
overall execution plan in parallel, each consumes a fraction of the
input. The chosen representation for each fragment, a.k.a. data
flow operation (DFO), is a partial view of the equivalent execution
plan on each thread, rather than a partial physical plan. So it is
critical to keep the overall plan equivalent on each thread,
including the coordinator and workers.

To generate equivalent plans, the optimizer must have exact same
inputs, for example SQL text, metadata, variables, statistics, and
so on.

Solutions:
=========
The first of the patch series supports copying variables that
are significant to optimization from the coordinator and reusing
them in workers. These variables include:

 - THD::user_vars
 - THD::variables
 - THD::table_plugin (default_storage_engine)
 - THD::temp_table_plugin (default_tmp_storage_engine)
 - THD::charset

MR!642 reviewed by kenkkchen & maxmqiu