[feature]#741 Allow getting the number of dynamic partitions
Problem:
=======

The PX scheduler might need the real number of dynamic partitions
to tune resource allocation, but it cannot get such information
with the current parallel scan interface. As a result, the scheduler
has to launch workers by configured DoP, which is a waste when it is
bigger than than the real number.

Solution:
========

Add the real number of partitions as an output param for
handler::ha_px_coordinator_init().

MR!639 reviewed by kenkkchen, maxmqiu