[bugfix]#718 Reserve ref_slice number for exchange
Problem:
========
The ref slice number for exchange tmp table is set to
REF_SLICE_WIN_1 + number of exchange. This conflicts with the
open end of the slice range reserved for window functions.

Solution:
=========
With the knowledge that the number of exchanges in a query block
is limited, reserve ref slice number for exchange in JOIN. The
reserved slices are REF_SLICE_EXCHANGE_{1,2,3} and the limit,
MAX_EXCHANGE_NUM, is checked in WalkAccessPathsForExchange.

MR!662 reviewed by kenkkchen maxmqiu