[feature]#578 Improve error handling for parallel execution
Problem:
========
Parallel execution (#406) had no cooperation between coordinator
and workers on errors. It could cause dedicated failure of
execution, missing error messages or deadly hang in abnormal cases:

1. The query coordinator might be killed by the user.
2. Any worker might be killed by the user.
3. Processing in any thread might return early on its own.

From the user's perspective, dedicated errors should be avoided as
many as possible, those inevitable should be clear, and the
overall execution should be responsive.

Solution:
=========
Currently parallel execution kicks in after regular optimization,
does some preparation work, executes tasks in parallel, writes to
the client user in the last task (a.k.a. root task), then does
cleanup.

If any error occurs before writing to the user or producing any
side effect, it has a chance to be recovered by retrying the
same query in a promising mode. For parallel execution, it is
serial execution. Queries are safe to be retried before writing
to the user, because they are read-only.

Obviously, only errors that could be recovered deserve the retrial.
They are errors dedicated to parallel execution. These errors are
retried, and others are reported to the user, wrapped in a new
error code, ER_PX_EXECUTE_ERROR.

Any error including the received KILL cause early return in the
thread. Any background worker reports to the coordinator, then the
coordinator broadcasts to all of the workers, so that each thread
is aware of the error. When all background workers join the
execution of the coordinator, the latter decides retrying the
query or just reporting to the user.

MR!613 reviewed by kenkkchen, maxmqiu