[bugfix]#776 Support compatibility check for union
Problem:
========
The current compatibility check design is at the statement level, and the compatibility
check results of different query blocks will affect each other. In a Union scenario, if
part of a Union Branch query block does not support parallelism, then the entire
statement will not run parallel execution, so we need to adjust the compatibility check
for this.

Solution:
=========
Each Query Block is checked for compatibility, and their results determine whether do
parallel execution for the entire statement. And union result with limit is refused.

MR!726 reviewed by kenkchen maxmqiu