[bugfix]#765 Inject exchange on SORT that is in the subtree of NestedLoopJoin
Problem:
========
In the one-stage parallelism plan, NestedLoopJoin and its
subsequent operators are generally pushed down into parallelism.
The only exception is when sort appears on the outer table of a
NestedLoopJoin, and only this branch with sort can be pushed down
into parallelism. The function that finds the inject position can
correctly determine where exchange should be injected, but because
Merge Sort was not supported previously, exchange is not actually
injected.

Solution:
=========
Insert Exchange on sort that is in the subtree of join, and use
Merge Receive for the final sort.

MR!687 reviewed by kenkchen, maxmqiu