[feature]#760 Support parallel stream aggregation with key order enforcer
Problem:
========
Parallel execution is based on the fact that the overall plan could
be divided into small pieces and some of them could be run in
parallel. A piece of plan is a.k.a. a fragment, a data flow
operation (DFO), or a stage.

The current version supports only one stage in parallel. And the
plan is roughly divided at the first pipeline breaker operator,
counting from the leaves (bottom) of the plan tree towards the root
(top).

There is a specific case that the key order of a steam aggregation
is enforced by a sorting operator. Such a plan is usually not
chosen by the cost-based optimizer (CBO) because of inefficiency
comparing to hash aggregation which eliminates the overhead of
sorting. However, things are changed when column store index (CSI)
could translate a piece into an extreamly efficient form that CBO
does not understand yet. So the specific case needs to be handled.

Solution:
=========
If a sorting operator serves as the key order enforcer of a stream
aggregation, divide the plan at the aggregation. The lower piece
is meant to be taken over by CSI.

MR!689 reviewed by kenkchen, maxmqiu