[feature]#713 Px_sender send the result of items
Problem:
========
Px_sender send the fields of table to the coordinator. If there is
an expression evaluation on some fields, the work on that
evaluation will be deferred to the coordinator.

Solution:
=========
Px_sender now can send the result of expressions.
Turn this feature on with the command
"set exchange_inject_use_item=on"

MR!617 by kenkkchen maxmqiu