[bugfix]#775 FindExchangeInjectPosition() misplaces exchange in some cases
Problem:
========
FindExchangeInjectPosition() traverses the access path tree in
pre-order by WalkAccessPaths() to find a node above which an
exchange operator will be inserted.

Previous implementation had some defects:

The traversal did not abort when any node was detected which did
not support parallel. Such a behavior was meaningless with parallel
compatibilty test at query block level. The traversal was also
incomplete in certain cases because the old mode,
STOP_AT_MATERIALIZATION, prunes below STREAM and MATERIALIZE.

In addition, STREAM and MATERIALIZE were falsely considered as
OK to be run in parallel.

A loosely related problem was that AGGREGATE might not be prepared
to run in parallel yet, i.e. split into local AGGREGATE and final
AGGREGATE, so exchange should not be inserted at all.

Solution:
=========
Abort traversal immediately when any node not supported yet is
found. Use ENTIRE_QUERY_BLOCK mode to do complete traversal for a
query block. ENTIRE_TREE might be considered in the future.

Note that when compatibility test is improved to be able to
identify a subtree as OK to run in parallel, the find function
needs to be reconsidered.

Detect AGGREGATE split failures and if any, make the plan serial.

In addition, removed HASH_JOIN and MATERIALIZE from the find
function because they are dead code due to commpatiblity test.

MR!705 reviewed by kenkchen, maxmqiu.