[bugfix] #949 Change semaphore to mutex and condvar
Problem:
========
MySQL KILL provides non-preemptive interrupt. It works by setting
the killed flag of target THD and awaking the thread to make it
responsive in case it is waiting. To be cooperative, the target
thread exposes its notification variable so as to be awaked,
and checks the flag and returns fast if it is set.

The general synchronization used in the previous implementation,
semaphore, did not support awaking as MySQL KILL expect. It would
require much code to fit the gap, which is also redundant.

Solution:
=========
Change semaphore to mutex and condition variable as MySQL KILL
expect.

Now the responsiveness of the target thread with respect to KILL
depends on its willingness to check the killed flag, which is left
to instrumenting the processing flow.

MR!920 reviewed by kenkkchen maxmqiu