[feature]#828 Decoupling of parallel optimization(2/2)
Problem:
========
Because the original parallel optimization framework is in the serial optimization
logic of a single QB, it is not suitable for the current compatibility check and
parallel optimization strategy.

Solution:
========
Replace the original parallel optimization framework with new compatibility check
and parallel optimization framework and clean up old code.

MR!881 reviewed by kenkkchen, maxmqiu