[feature]#858 Add a private cache for statistics requests
Problem
=======
MySQL ndependently requests statistics for each reference to the
same physical table and updates certain shared statistics (See also
Bug#106264), which both adds overhead and makes the view of
statistics inconsistent. The former is notable in some cases, and
the latter was reported to cause bad plans with a wrong index.

Solution
========
For now a private cache is introduced between the optimizer and
the statistics component. The result of each statistics request,
records_in_range() or info(), is cached for further reuse.

A sysvar is added to turn it on:
 - cdb_optimization_context_cache_enabled (on/off)

The cache involves these statistics:
  - records in range (index dive)
  - handler:stats (table rows and other table-level statistics)
  - TABLE_SHARE::key_info::rec_per_key/rec_per_key_float
  - in_memory_estimate

Column statistics are not involved due to protection by MDL lock.

Statistics inconsistency is thus solved by the private cache. Note
that MySQL might also perform execution in the optimization phase,
which retrieves data instead of statistics, thus is irrelevant.
Data consistency is enforced by read view.

Notes
=====
Range estimate might still be inconsistent because index dive does
not understand ranges over multiple indexes. It is to be fixed in
the future. A shared cache may be added to further reduce overhead.

Since the cache is active during the lifetime of the top statement,
it might explode because of functions. Index dive for certain
conditions might also consume a notable amount of memory. Memory
limit might need to be considered in the future.

MR!767 reviewed by kenkkchen & maxmqiu & casazhang